check dependencies moved to the scripts folder
